print speed leve.

+++ src/aliengo_bridge.cpp
@@ -251,20 +251,20 @@ namespace aliengo_bridge
             else if (((int)m_joy.btn.components.Y == 1) && m_cmd_speed_level < 2)
             {
                 m_cmd_speed_level += 1;
-                ROS_INFO_STREAM("Increased speed level by 1 to: " << m_cmd_speed_level);
+                ROS_INFO_STREAM("Increased speed level by 1 to: " << (int)m_cmd_speed_level);
                 m_last_velocity_update = ros::Time::now();
             }
             else if (((int)m_joy.btn.components.A == 1) && m_cmd_speed_level > 0)
             {
                 m_cmd_speed_level -= 1;
-                ROS_INFO_STREAM("Decreased speed level by 1 to: " << m_cmd_speed_level);
+                ROS_INFO_STREAM("Decreased speed level by 1 to: " << (int)m_cmd_speed_level);
                 m_last_velocity_update = ros::Time::now();
             }
             else if (((int)m_joy.btn.components.F1 == 1))
             {
                 m_cmd_speed_level = 0;
                 m_cmd_max_velocity = 0.0f;
-                ROS_INFO_STREAM("Reset max velocity and speed level to: " << m_cmd_max_velocity << " and " << m_cmd_speed_level);
+                ROS_INFO_STREAM("Reset max velocity and speed level to: " << m_cmd_max_velocity << " and " << (int)m_cmd_speed_level);
                 m_last_velocity_update = ros::Time::now();
             }
             else if (((int)m_joy.btn.components.select == 1))
